Refactored the Renderer::Submit function
Refactored the Renderer Submit function, added returns for the copy constructors of assets if they were invalid. Cleaned up the scene update function to look better and work with the new submit call. Also changed material struct to use XMFLOAT4 and not XMVECTOR and added a pad to match the new change.