Render textures and cubemaps working now
there is no evidence but TRUST ME BRO. anyway, you can render into a render cubemap and use it as an input for a later pass. and the render graph would(probably) transition it. another cool thing would be to bind the entire cubemap as an rgtexture and have all arrays transitioned.